+++ backend/onyx/agents/agent_search/deep_search_a/initial/generate_sub_answers/graph_builder.py
@@ -2,23 +2,23 @@ from langgraph.graph import END
 from langgraph.graph import START
 from langgraph.graph import StateGraph
 
-from onyx.agents.agent_search.deep_search_a.initial.consolidate_sub_answers.edges import (
-    parallelize_initial_sub_question_answering,
-)
-from onyx.agents.agent_search.deep_search_a.initial.consolidate_sub_answers.nodes.decompose_orig_question import (
-    decompose_orig_question,
-)
-from onyx.agents.agent_search.deep_search_a.initial.consolidate_sub_answers.nodes.format_initial_sub_answers import (
-    format_initial_sub_answers,
-)
-from onyx.agents.agent_search.deep_search_a.initial.consolidate_sub_answers.states import (
-    SQInput,
-)
-from onyx.agents.agent_search.deep_search_a.initial.consolidate_sub_answers.states import (
-    SQState,
-)
 from onyx.agents.agent_search.deep_search_a.initial.generate_individual_sub_answer.graph_builder import (
     answer_query_graph_builder,
+)
+from onyx.agents.agent_search.deep_search_a.initial.generate_sub_answers.edges import (
+    parallelize_initial_sub_question_answering,
+)
+from onyx.agents.agent_search.deep_search_a.initial.generate_sub_answers.nodes.decompose_orig_question import (
+    decompose_orig_question,
+)
+from onyx.agents.agent_search.deep_search_a.initial.generate_sub_answers.nodes.format_initial_sub_answers import (
+    format_initial_sub_answers,
+)
+from onyx.agents.agent_search.deep_search_a.initial.generate_sub_answers.states import (
+    SQInput,
+)
+from onyx.agents.agent_search.deep_search_a.initial.generate_sub_answers.states import (
+    SQState,
 )
 from onyx.utils.logger import setup_logger
 
@@ -27,7 +27,7 @@ logger = setup_logger()
 test_mode = False
 
 
-def consolidate_sub_answers_graph_builder() -> StateGraph:
+def generate_sub_answers_graph_builder() -> StateGraph:
     graph = StateGraph(
         state_schema=SQState,
         input=SQInput,
@@ -37,10 +37,10 @@ def consolidate_sub_answers_graph_builder() -> StateGraph:
         node="decompose_orig_question",
         action=decompose_orig_question,
     )
-    answer_sub_question_subgraph = answer_query_graph_builder().compile()
+    answer_sub_question_subgraphs = answer_query_graph_builder().compile()
     graph.add_node(
-        node="answer_sub_question_subgraph",
-        action=answer_sub_question_subgraph,
+        node="answer_sub_question_subgraphs",
+        action=answer_sub_question_subgraphs,
     )
 
     graph.add_node(
@@ -75,10 +75,10 @@ def consolidate_sub_answers_graph_builder() -> StateGraph:
     graph.add_conditional_edges(
         source="decompose_orig_question",
         path=parallelize_initial_sub_question_answering,
-        path_map=["answer_sub_question_subgraph"],
+        path_map=["answer_sub_question_subgraphs"],
     )
     graph.add_edge(
-        start_key="answer_sub_question_subgraph",
+        start_key=["answer_sub_question_subgraphs"],
         end_key="format_initial_sub_question_answers",
     )
